requirements.txt
streamlit==1.42.2
pandas==2.2.3
psycopg2-binary==2.9.9
altair==5.2.0
jira==3.5.1